requirements.txt
torch
torchvision
opencv-python
numpy
scikit-learn
matplotlib
pyserial
Pillow
luma.oled